modules/gguf_model.py
# gguf_model.py

import os
import torch
from llama_cpp import Llama

class GGUFModel:
    def __init__(self, model_path, n_ctx=512, n_gpu_layers=0, n_threads=4, **kwargs):
        """
        Initialize the GGUF model with specified parameters.

        :param model_path: Path to the GGUF model file.
        :param n_ctx: Context length.
        :param n_gpu_layers: Number of layers to offload to GPU.
        :param n_threads: Number of CPU threads to use.
        :param kwargs: Additional parameters.
        """
        self.model_path = model_path
        self.n_ctx = n_ctx
        self.n_threads = n_threads
        self.kwargs = kwargs
        self.model = None

        # Detect CUDA availability
        self.device = "cuda" if torch.cuda.is_available() else "cpu"
        self.n_gpu_layers = n_gpu_layers if self.device == "cuda" else 0

    def load_model(self):
        """
        Load the GGUF model using llama_cpp Python bindings.
        """
        if not os.path.exists(self.model_path):
            raise FileNotFoundError(f"Model file {self.model_path} not found.")
        self.model = Llama(
            model_path=self.model_path,
            n_ctx=self.n_ctx,
            n_gpu_layers=self.n_gpu_layers,
            n_threads=self.n_threads,
            **self.kwargs
        )

    def unload_model(self):
        """
        Unload the GGUF model to free up resources.
        """
        if self.model:
            del self.model
            self.model = None

    def generate(self, prompt, max_tokens=256, stop=None):
        """
        Generate text based on the input prompt.

        :param prompt: Input text prompt.
        :param max_tokens: Maximum number of tokens to generate.
        :param stop: Sequence of tokens to stop generation.
        :return: Generated text.
        """
        if self.model is None:
            raise ValueError("Model is not loaded.")
        output = self.model(prompt, max_tokens=max_tokens, stop=stop)
        return output["choices"][0]["text"]

    def get_attention_weights(self, input_ids):
        """
        Retrieve attention weights from the model (if supported).
        """
        # Placeholder for actual implementation
        pass

    def get_weights(self):
        """
        Retrieve model weights for visualization.
        """
        # Placeholder for actual implementation
        pass

    def get_embeddings(self, input_ids):
        """
        Retrieve node embeddings from the model.
        """
        # Placeholder for actual implementation
        pass
